scraper/requirements.txt
requests
playwright
python-dotenv
duckduckgo-search
